requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
coverage==5.5
crayons==0.4.0
cryptography==3.4.7
Django==3.2.3
django-alphafilter==1.0
django-environ==0.4.5
django-extensions==3.1.3
django-rest-knox==4.1.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
fake-useragent==0.1.11
flake8==3.9.2
future==0.18.2
graphviz==0.16
idna==2.10
lxml==4.6.3
lyricsgenius==3.0.1
mccabe==0.6.1
numpy==1.21.0
openpyxl==3.0.7
pandas==1.3.0
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tqdm==4.61.2
urllib3==1.26.5
webdriver-manager==3.4.2
XlsxWriter==1.4.4
